Ph calculation sure the remote path is there (#115)
* Ph calculation sure the remote path is there

The caching mechanism is tricky for the phonon calculation.
See discussion https://github.com/aiidateam/aiida-core/issues/5178
It is workaround by before running ph calculation, always check the
remote path of the previous pw calculation. If it is empty means the
folder is cleaned therefore potentially will fail the PH calculation.
And the pw calculation will actually run without from using caching
node.

It is not a perfect solution since even if the pw calculation workdir is
being cleaned, if the subseqent ph calculation is succeed and used as
caching node, in principle the pw calculation no need to run from
scratch without caching.

also include:
- fix: propagate if sub-workflow of verfi workflow not finished_ok
- Unify the properties name seprate with dot.
- fix: datatype issue of accuracy.delta and accuracy.bands

+++ aiida_sssp_workflow/workflows/measure/delta.py
@@ -188,8 +188,8 @@ class DeltaMeasureWorkChain(SelfCleanWorkChain):
             "element": orm.Str(self.ctx.element),  # _base wf hold attribute `element`
             "configuration": orm.Str(configuration),
             "pw_base_parameters": orm.Dict(dict=self.ctx.pw_parameters),
-            "ecutwfc": orm.Float(self.ctx.ecutwfc),
-            "ecutrho": orm.Float(self.ctx.ecutrho),
+            "ecutwfc": orm.Int(self.ctx.ecutwfc),
+            "ecutrho": orm.Int(self.ctx.ecutrho),
             "kpoints_distance": orm.Float(self.ctx.kpoints_distance),
             "scale_count": orm.Int(self.ctx.scale_count),
             "scale_increment": orm.Float(self.ctx.scale_increment),
